requirements.txt
appdirs==1.4.4
botocore==1.20.21
certifi==2020.12.5
colorama==0.4.4
configparser==5.0.2
distlib==0.3.1
dnspython==2.1.0
filelock==3.0.12
grpc-interceptor==0.13.0
grpcio==1.36.1
grpcio-health-checking==1.36.1
grpcio-tools==1.36.1
jmespath==0.10.0
marshmallow==3.10.0
minio==6.0.0
mongoengine==0.23.0
packaging==20.9
pluggy==0.13.1
protobuf==3.15.5
py==1.10.0
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.15.0
toml==0.10.2
tox==3.23.0
urllib3==1.26.3
virtualenv==20.4.2
